[flutter_tools] move some fakes out of mocks.dart (#74992)
