+++ packages/flutter_tools/test/commands.shard/hermetic/shell_completion_test.dart
@@ -16,33 +16,33 @@ import 'package:flutter_tools/src/globals.dart' as globals;
 
 import '../../src/common.dart';
 import '../../src/context.dart';
-import '../../src/mocks.dart';
+import '../../src/fakes.dart';
 
 void main() {
   group('shell_completion', () {
-    MockStdio mockStdio;
+    FakeStdio fakeStdio;
 
     setUp(() {
       Cache.disableLocking();
-      mockStdio = MockStdio()..stdout.terminalColumns = 80;
+      fakeStdio = FakeStdio()..stdout.terminalColumns = 80;
     });
 
     testUsingContext('generates bash initialization script to stdout', () async {
       final ShellCompletionCommand command = ShellCompletionCommand();
       await createTestCommandRunner(command).run(<String>['bash-completion']);
-      expect(mockStdio.writtenToStdout.length, equals(1));
-      expect(mockStdio.writtenToStdout.first, contains('__flutter_completion'));
+      expect(fakeStdio.writtenToStdout.length, equals(1));
+      expect(fakeStdio.writtenToStdout.first, contains('__flutter_completion'));
     }, overrides: <Type, Generator>{
-      Stdio: () => mockStdio,
+      Stdio: () => fakeStdio,
     });
 
     testUsingContext('generates bash initialization script to stdout with arg', () async {
       final ShellCompletionCommand command = ShellCompletionCommand();
       await createTestCommandRunner(command).run(<String>['bash-completion', '-']);
-      expect(mockStdio.writtenToStdout.length, equals(1));
-      expect(mockStdio.writtenToStdout.first, contains('__flutter_completion'));
+      expect(fakeStdio.writtenToStdout.length, equals(1));
+      expect(fakeStdio.writtenToStdout.first, contains('__flutter_completion'));
     }, overrides: <Type, Generator>{
-      Stdio: () => mockStdio,
+      Stdio: () => fakeStdio,
     });
 
     testUsingContext('generates bash initialization script to output file', () async {
@@ -56,7 +56,7 @@ void main() {
     }, overrides: <Type, Generator>{
       FileSystem: () => MemoryFileSystem.test(),
       ProcessManager: () => FakeProcessManager.any(),
-      Stdio: () => mockStdio,
+      Stdio: () => fakeStdio,
     });
 
     testUsingContext("won't overwrite existing output file ", () async {
@@ -77,7 +77,7 @@ void main() {
     }, overrides: <Type, Generator>{
       FileSystem: () => MemoryFileSystem.test(),
       ProcessManager: () => FakeProcessManager.any(),
-      Stdio: () => mockStdio,
+      Stdio: () => fakeStdio,
     });
 
     testUsingContext('will overwrite existing output file if given --overwrite', () async {
@@ -92,7 +92,7 @@ void main() {
     }, overrides: <Type, Generator>{
       FileSystem: () => MemoryFileSystem.test(),
       ProcessManager: () => FakeProcessManager.any(),
-      Stdio: () => mockStdio,
+      Stdio: () => fakeStdio,
     });
   });
 }
