+++ packages/flutter_tools/test/general.shard/analytics_test.dart
@@ -26,7 +26,7 @@ import 'package:usage/usage_io.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
-import '../src/mocks.dart';
+import '../src/fakes.dart';
 
 void main() {
   setUpAll(() {
@@ -152,14 +152,14 @@ void main() {
 
   group('analytics with mocks', () {
     MemoryFileSystem memoryFileSystem;
-    MockStdio mockStdio;
+    FakeStdio fakeStdio;
     TestUsage testUsage;
     FakeClock fakeClock;
     Doctor mockDoctor;
 
     setUp(() {
       memoryFileSystem = MemoryFileSystem.test();
-      mockStdio = MockStdio();
+      fakeStdio = FakeStdio();
       testUsage = TestUsage();
       fakeClock = FakeClock();
       mockDoctor = MockDoctor();
@@ -251,7 +251,7 @@ void main() {
           'FLUTTER_ANALYTICS_LOG_FILE': 'analytics.log',
         },
       ),
-      Stdio: () => mockStdio,
+      Stdio: () => fakeStdio,
     });
 
     testUsingContext('event sends localtime', () async {
@@ -281,7 +281,7 @@ void main() {
           'FLUTTER_ANALYTICS_LOG_FILE': 'analytics.log',
         },
       ),
-      Stdio: () => mockStdio,
+      Stdio: () => fakeStdio,
     });
   });
