+++ packages/flutter_tools/test/general.shard/android/android_workflow_test.dart
@@ -18,7 +18,8 @@ import 'package:mockito/mockito.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
-import '../../src/mocks.dart' show MockAndroidSdk, MockProcess, MockProcessManager, MockStdio;
+import '../../src/fakes.dart';
+import '../../src/mocks.dart' show MockAndroidSdk, MockProcess, MockProcessManager;
 import '../../src/testbed.dart';
 
 class MockAndroidSdkVersion extends Mock implements AndroidSdkVersion {}
@@ -29,7 +30,7 @@ void main() {
   Logger logger;
   MemoryFileSystem fileSystem;
   MockProcessManager processManager;
-  MockStdio stdio;
+  FakeStdio stdio;
 
   setUp(() {
     sdk = MockAndroidSdk();
@@ -37,7 +38,7 @@ void main() {
     fileSystem.directory('/home/me').createSync(recursive: true);
     logger = BufferLogger.test();
     processManager = MockProcessManager();
-    stdio = MockStdio();
+    stdio = FakeStdio();
   });
 
   MockProcess Function(List<String>) processMetaFactory(List<String> stdout) {
