+++ packages/flutter_tools/test/general.shard/base/bot_detector_test.dart
@@ -16,12 +16,12 @@ import 'package:mockito/mockito.dart';
 import 'package:fake_async/fake_async.dart';
 
 import '../../src/common.dart';
-import '../../src/mocks.dart';
+import '../../src/fakes.dart';
 
 void main() {
   group('BotDetector', () {
     FakePlatform fakePlatform;
-    MockStdio mockStdio;
+    FakeStdio fakeStdio;
     MockHttpClient mockHttpClient;
     MockHttpClientRequest mockHttpClientRequest;
     MockHttpHeaders mockHttpHeaders;
@@ -30,7 +30,7 @@ void main() {
 
     setUp(() {
       fakePlatform = FakePlatform()..environment = <String, String>{};
-      mockStdio = MockStdio();
+      fakeStdio = FakeStdio();
       mockHttpClient = MockHttpClient();
       mockHttpClientRequest = MockHttpClientRequest();
       mockHttpHeaders = MockHttpHeaders();
@@ -66,9 +66,9 @@ void main() {
         when(mockHttpClient.getUrl(any)).thenAnswer((_) {
           throw const SocketException('HTTP connection timed out');
         });
-        mockStdio.stdout.hasTerminal = true;
+        fakeStdio.stdout.hasTerminal = true;
         expect(await botDetector.isRunningOnBot, isFalse);
-        mockStdio.stdout.hasTerminal = false;
+        fakeStdio.stdout.hasTerminal = false;
         expect(await botDetector.isRunningOnBot, isFalse);
         expect(persistentToolState.isRunningOnBot, isFalse);
       });
