+++ packages/flutter_tools/test/general.shard/device_test.dart
@@ -18,6 +18,7 @@ import 'package:fake_async/fake_async.dart';
 import '../src/common.dart';
 import '../src/context.dart';
 import '../src/fake_devices.dart';
+import '../src/fakes.dart';
 import '../src/mocks.dart';
 
 void main() {
